fix(projects): make assignment fetch relay-real and partial-publish heal reachable
Round-3 review (wpfleger96) found two defects living in contracts the
previous tests abstracted away:

1. Assignment-operation fetch re-created the bounded window it was meant
   to fix. The relay pushes only kinds/ids/authors/#e/single-#p/NIP-33-#d
   into SQL; #t/#a are post-filtered in Rust AFTER the SQL LIMIT
   (filter_to_query_params / filter_fully_pushable in
   crates/buzz-relay/src/handlers/req.rs). A #t-carrying filter therefore
   returns the newest 500 kind-1 candidates tag-filtered down — a short
   (often empty) page falsely signals exhaustion and an assignment older
   than ~500 unrelated notes is silently lost.

   The fetch is now keyed per issue via #e (SQL-pushed JSONB containment,
   issue ids already in hand at both call sites) with assignment labels
   filtered client-side, so a short page is a true end-of-results signal.
   The same-second edge (a full page whose oldest timestamp equals the
   inclusive until cursor) widens once to the relay's 1,000-row clamp and
   errors beyond that — surfacing a failed assignments section instead of
   silent loss, since NIP-01 filters cannot express the relay's composite
   (created_at, id) keyset cursor. Regression stubs now model the
   production relay: SQL-pushed constraints, LIMIT, then Rust
   post-filtering — plus the 0/1 buried-assignment case and both
   same-second densities.

2. The resume-aware dominated-write guard was unreachable for the very
   retry it exists to allow: after a partial publish the live head is
   necessarily newer than the cached snapshot (the failed mutation never
   updated the cache), so the earlier unconditional guard threw
   'updated by another session' before the dangling-member probe ran —
   breaking the error message's own 're-run Add repository to resume'
   promise. The unconditional guard is deleted; the resume-aware guard
   provides identical protection for the non-resume case. The regression
   now exercises the exported mutation itself (guards included) through
   injected relay/publish seams: heal on both local and owner-agent
   paths, dominated-write still throws without a dangling member, and
   resume works with an unmoved head.
