+++ desktop/src/shared/ui/mentionChip.ts
@@ -8,6 +8,7 @@ export type InlineChipIconKind =
   | "channel"
   | "message"
   | "repo"
+  | "project"
   | "pr"
   | "issue";
 
@@ -17,6 +18,7 @@ const INLINE_CHIP_ICON_KIND_CLASSES: Record<InlineChipIconKind, string> = {
   channel: "inline-chip-icon-channel",
   message: "inline-chip-icon-message",
   repo: "inline-chip-icon-repo",
+  project: "inline-chip-icon-project",
   pr: "inline-chip-icon-pr",
   issue: "inline-chip-icon-issue",
 };
